+++ src/index.ts
@@ -2,7 +2,7 @@ import { getWaybackItemsWithLocalChanges } from './change-detector';
 import { getMetadata } from './metadata';
 import { getWaybackItems } from './wayback-items';
 import { WaybackItem, WaybackMetadata, WaybackConfig } from './types';
-import { setDefaultWaybackOptions } from './config';
+import { setDefaultWaybackOptions, getWaybackSubDomains } from './config';
 
 import { long2tile, lat2tile, tile2Long, tile2lat } from './helpers/geometry';
 
@@ -11,6 +11,7 @@ export {
     getWaybackItems,
     getMetadata,
     setDefaultWaybackOptions,
+    getWaybackSubDomains,
     long2tile,
     lat2tile,
     tile2Long,